requirements.txt
streamlit
pandas
numpy
plotly
pdfkit
jinja2 
pytesseract
pillow
regex
matplotlib
reportlab

